Start the creation of the Output class.
The code will be more clean and more easily to change. Idea was
borne when trying to implement a nospace / nocapital option. There
was no easy way to make it optional.

The class is not yet used. It will replace a few parts of other
classes. It's already pretty complete, though the dir_exists notice
still has to handled properly. All the necessary functions should
be there though.

After having fixed that, the big task will be to let all other code
use this new first class citizen.