Fix a problem for certain drives with cdparanoia and the last track.
The last track is now passed as a '-' meaning till the end of the
disc. I suspect that cdparanoia will behave nicely once again on
certain problematic drives. A few issues were raised because
cdparanoia was giving problems with the last track.

It even led some people to conclude that rubyripper was not
passing the correct values. That was not true. But I think
I fixed it anyhow as a nice workaround ;)