Fix a problem that only showed up in the command line client.
Though possibly it would give some problems with the gui client
as well. Ruby was deleting objects it though weren't used anymore.
But that was not true.  I still think it's sort of a ruby bug.

Anyway, now a more simple, chronological communication from the
frontend to the backend exists which solved this problem.