First actual workable implementation of the AdvancedToc class.
Now your disc is by default scanned before being ripped. This
doesn't slow down your ripping so much as you might expect. It
seems that cdrdao is doing a warming-up for your drive to enable
it to have excellent speed.

The information of the toc file of cdrdao is only interpreted now.
It is shown in the gui as a log message. Nothing is done with it
yet. That would be the next step. The different solutions are not
that clear right now. But that may soon change.