Make the trackname a Hash variable. This allows for setting track 0.
I've tried a hack to do this, but this didn't work out nicely. The
more clean way is to actually support a track 0 natively, so all
other classes still work as expected. A surprising few lines were
needed to change for this ;)

Off course the trackname still has to be filled somewhere.