First restyling of the Disc class to help functions.
Try to hide the implementation for the other classes. Real object
oriented means that each class should handle it's own calculations.
In the future the Disc class should be able to handle the input
that cdparanoia needs for all sectors. I tend to let AdvancedToc
update the Disc class with more detailed info. This way the rest
of the program doesn't need to be bothered to check the AdvancedToc
class.

Also change the way image rips are done. Doing this by calling it
track 0 is just a dirty hack. Besides it is conflicting with the
way rubyripper tries to handle pregaps for the first track. This
audio is also marked as track 0.