Probably fix the problem in issue 362.
Before undo marking the disc as various artist, first check if the
current disc has various artist set. It just might be that this was
only the case with the previous disc. This resulted in a crash for
the gui. It was feed with an empty Array instead of a String Array.
The backup info of varArtist info was reset with the disc update.