Give a warning instead of an endless loop when offset > 0.
When offset > 0 some drives don't properly output the right size.
Instead they chop off a few sectors at the end. Notice that this
may very well be a limitation of cdparanoia. As alternative
behaviour the amount of sectors is reported, also when the amount of
sectors is not an integer value.