Refactor the Disc and FreedbString class.
With the new Cdcontrol parser it was getting really ugly the way
it was. So clean it up. FreedbString class should be all about
creating the string, not managing the right scanner.

The Disc class now properly forwards requests with a minimum line
of codes needed. Also the dependency creation has been moved to
the child classes, which improves readability quite nicely.

In the process also added a rspec file for the Disc class. Of
course the other specs are cleaned as well.