First attempt fixing translation. Please read the extra info.
This is a two-way process.
1) Make sure to bind all relevant classes to gettext.
2) Fixing the configure file to create new po/mo files.

So far I've fixed all specs + cucumber tests. The rest will crash
probably, all classes that use translation should add the following
code at top (see example in committed files):

* include GetText
* GetText.bindtextdomain("rubyripper")

When the _() translate function is called before the initalize takes
place, it is necessary to add a third line:
* def self._(txt) ; GetText._(txt) ; end

Because I don't want all kind of messages spitting through the console
I fixed the specs and cucumber to be silent. I did this as follows:

Rspec: In the env file I've defined the fake gettext class directly.
Cucumber: I've added an ENV variable while running. This should be the
default for all cucumber detect logic.

A complication exists for gettext in that it is currently crashing in
all kinds of weird ways. This has to be fixed upstream. I trust this
won't take long.