Did some analysis on the performance of the ripping.
First suspect was the CRC calculation, but this turned out to be
wrong. Just as the MD5 calculation it is quite fast. They are now
calculated together within 1 second for a 2 minute track on my
machine. This is reasonable.

What is not, is the peak calculation which takes about 25 seconds
for the same track. Do we even need this? Can't we optimize this
beast somehow or make it optional?